+++ skis-test-model/src/main/java/io/skis/entity/Department.java
@@ -1,0 +1,69 @@
+package io.skis.entity;
+
+import io.skis.annotations.Column;
+import io.skis.annotations.Id;
+import io.skis.annotations.SkisEntity;
+import io.skis.annotations.Table;
+import io.skis.annotations.Version;
+import java.time.Instant;
+
+/**
+ * @author: Hu Xin
+ */
+@SkisEntity
+@Table(name = "CL_DEPARTMENT")
+public class Department {
+
+  @Id private Long id;
+
+  @Column(nullable = false)
+  private Instant createStamp;
+
+  @Column(nullable = false)
+  private Instant modifyStamp;
+
+  @Version private Long version;
+
+  @Column(nullable = false)
+  private String name;
+
+  public Long getId() {
+    return id;
+  }
+
+  public void setId(Long id) {
+    this.id = id;
+  }
+
+  public Instant getCreateStamp() {
+    return createStamp;
+  }
+
+  public void setCreateStamp(Instant createStamp) {
+    this.createStamp = createStamp;
+  }
+
+  public Instant getModifyStamp() {
+    return modifyStamp;
+  }
+
+  public void setModifyStamp(Instant modifyStamp) {
+    this.modifyStamp = modifyStamp;
+  }
+
+  public Long getVersion() {
+    return version;
+  }
+
+  public void setVersion(Long version) {
+    this.version = version;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+}